+++ core-ng/src/test/java/core/framework/impl/db/RepositoryImplAssignedIdEntityTest.java
@@ -14,8 +14,9 @@ import java.time.Month;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 /**
@@ -31,6 +32,7 @@ class RepositoryImplAssignedIdEntityTest {
         database = new DatabaseImpl("db");
         database.url("jdbc:hsqldb:mem:.;sql.syntax_mys=true");
         database.vendor = Vendor.MYSQL;
+        database.operation.batchSize = 7;
         database.execute("CREATE TABLE assigned_id_entity (id VARCHAR(36) PRIMARY KEY, string_field VARCHAR(20), int_field INT, big_decimal_field DECIMAL(10,2), date_field DATE)");
 
         repository = database.repository(AssignedIdEntity.class);
@@ -48,31 +50,19 @@ class RepositoryImplAssignedIdEntityTest {
 
     @Test
     void insert() {
-        AssignedIdEntity entity = new AssignedIdEntity();
-        entity.id = UUID.randomUUID().toString();
-        entity.stringField = "string";
-        entity.intField = 12;
+        AssignedIdEntity entity = entity(UUID.randomUUID().toString(), "string", 12);
         entity.bigDecimalField = new BigDecimal("86.99");
         entity.dateField = LocalDate.of(2016, Month.JULY, 5);
 
         Optional<Long> id = repository.insert(entity);
         assertFalse(id.isPresent());
 
-        AssignedIdEntity selectedEntity = repository.get(entity.id).orElseThrow();
-
-        assertEquals(entity.id, selectedEntity.id);
-        assertEquals(entity.stringField, selectedEntity.stringField);
-        assertEquals(entity.intField, selectedEntity.intField);
-        assertEquals(entity.bigDecimalField, selectedEntity.bigDecimalField);
-        assertEquals(entity.dateField, selectedEntity.dateField);
+        assertThat(repository.get(entity.id)).get().isEqualToComparingFieldByField(entity);
     }
 
     @Test
     void update() {
-        AssignedIdEntity entity = new AssignedIdEntity();
-        entity.id = UUID.randomUUID().toString();
-        entity.stringField = "string";
-        entity.intField = 11;
+        AssignedIdEntity entity = entity(UUID.randomUUID().toString(), "string", 11);
         repository.insert(entity);
 
         AssignedIdEntity updatedEntity = new AssignedIdEntity();
@@ -81,59 +71,52 @@ class RepositoryImplAssignedIdEntityTest {
         updatedEntity.dateField = LocalDate.of(2016, Month.JULY, 5);
         repository.update(updatedEntity);
 
-        AssignedIdEntity selectedEntity = repository.get(entity.id).orElseThrow();
-        assertEquals(updatedEntity.stringField, selectedEntity.stringField);
-        assertEquals(entity.intField, selectedEntity.intField);
-        assertEquals(updatedEntity.dateField, selectedEntity.dateField);
+        assertThat(repository.get(entity.id)).get().isEqualToComparingOnlyGivenFields(updatedEntity, "stringField", "dateField");
     }
 
     @Test
     void delete() {
-        AssignedIdEntity entity = new AssignedIdEntity();
-        entity.id = UUID.randomUUID().toString();
-        entity.intField = 11;
+        AssignedIdEntity entity = entity(UUID.randomUUID().toString(), "string", 11);
         repository.insert(entity);
 
         repository.delete(entity.id);
 
-        Optional<AssignedIdEntity> result = repository.get(entity.id);
-        assertFalse(result.isPresent());
+        assertThat(repository.get(entity.id)).isNotPresent();
     }
 
     @Test
     void batchInsert() {
-        AssignedIdEntity entity1 = new AssignedIdEntity();
-        entity1.id = "1";
-        entity1.stringField = "value1";
-        entity1.intField = 11;
+        List<AssignedIdEntity> entities = Lists.newArrayList();
+        for (int i = 1; i < 100; i++) {
+            AssignedIdEntity entity = entity(String.valueOf(i), "value" + i, 10 + i);
+            entities.add(entity);
+        }
+        repository.batchInsert(entities);
 
-        AssignedIdEntity entity2 = new AssignedIdEntity();
-        entity2.id = "2";
-        entity2.stringField = "value2";
-        entity2.intField = 12;
-
-        repository.batchInsert(List.of(entity1, entity2));
-
-        AssignedIdEntity selectedEntity1 = repository.get("1").orElseThrow();
-        assertEquals(entity1.stringField, selectedEntity1.stringField);
-
-        AssignedIdEntity selectedEntity2 = repository.get("2").orElseThrow();
-        assertEquals(entity2.stringField, selectedEntity2.stringField);
+        assertThat(repository.get("1")).get().isEqualToComparingFieldByFieldRecursively(entities.get(0));
+        assertThat(repository.get("2")).get().isEqualToComparingFieldByFieldRecursively(entities.get(1));
     }
 
     @Test
     void batchDelete() {
-        AssignedIdEntity entity1 = new AssignedIdEntity();
-        entity1.id = "3";
-        entity1.intField = 11;
-        AssignedIdEntity entity2 = new AssignedIdEntity();
-        entity2.id = "4";
-        entity2.intField = 12;
-        repository.batchInsert(Lists.newArrayList(entity1, entity2));
+        List<AssignedIdEntity> entities = Lists.newArrayList();
+        for (int i = 100; i < 200; i++) {
+            AssignedIdEntity entity = entity(String.valueOf(i), "value" + i, 10 + i);
+            entities.add(entity);
+        }
+        repository.batchInsert(entities);
 
-        repository.batchDelete(Lists.newArrayList(entity1.id, entity2.id));
+        repository.batchDelete(entities.stream().map(entity -> entity.id).collect(Collectors.toList()));
 
-        assertFalse(repository.get(entity1.id).isPresent());
-        assertFalse(repository.get(entity2.id).isPresent());
+        assertThat(repository.get(entities.get(0).id)).isNotPresent();
+        assertThat(repository.get(entities.get(1).id)).isNotPresent();
+    }
+
+    private AssignedIdEntity entity(String id, String stringField, int intField) {
+        AssignedIdEntity entity = new AssignedIdEntity();
+        entity.id = id;
+        entity.stringField = stringField;
+        entity.intField = intField;
+        return entity;
     }
 }